+++ src/math/backends/webgl/clip_gpu.ts
