+++ src/math/backends/webgl/multinomial_gpu.ts
