+++ src/math/backends/webgl/texture_manager.ts
