+++ src/math/types.ts
@@ -1,0 +1,11 @@
+export interface SumTypes {
+  float32: 'float32';
+  int32: 'int32';
+  bool: 'int32';
+}
+
+export enum SumTypesMap {
+  float32 = 'float32',
+  int32 = 'int32',
+  bool = 'int32'
+}
