+++ src/test_util.ts
@@ -17,9 +17,7 @@
 
 import * as environment from './environment';
 import {Environment, Features} from './environment';
-import {NDArrayMath} from './math/math';
-import {NDArrayMathCPU} from './math/math_cpu';
-import {NDArrayMathGPU} from './math/math_gpu';
+import {NDArrayMath, NDArrayMathCPU, NDArrayMathGPU} from './math/math';
 import * as util from './util';
 import {DType, TypedArray} from './util';
 
@@ -201,7 +199,7 @@ export type Tests = (it: (name: string, testFn: () => void) => void) => void;
 
 export function describeMathCPU(
     name: string, tests: MathTests[], featuresList?: Features[]) {
-  const testNameBase = 'math_cpu.' + name;
+  const testNameBase = 'CPU: math.' + name;
   describeWithFeaturesAndExecutor(
       testNameBase, tests as Tests[],
       (testName, tests, features) => executeMathTests(
@@ -211,7 +209,7 @@ export function describeMathCPU(
 
 export function describeMathGPU(
     name: string, tests: MathTests[], featuresList?: Features[]) {
-  const testNameBase = 'math_gpu.' + name;
+  const testNameBase = 'WebGL: math.' + name;
   describeWithFeaturesAndExecutor(
       testNameBase, tests as Tests[],
       (testName, tests, features) => executeMathTests(
